+++ app/src/main/java/com/teamtriad/forpets/ui/adopt/AdoptFragment.kt
@@ -9,7 +9,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import com.squareup.moshi.JsonDataException
 import com.teamtriad.forpets.AdoptService
-import com.teamtriad.forpets.adapter.AdoptAdapter
+import com.teamtriad.forpets.ui.adopt.adapter.AdoptAdapter
 import com.teamtriad.forpets.databinding.FragmentAdoptBinding
 import com.teamtriad.forpets.model.AbandonmentInfo
 import com.teamtriad.forpets.util.toYyyyMmDd
